Fixes in parsing and updates to unity 2017.4.29

+++ Assets/uAdventure/Editor/Engine logic/Controllers/Controller.cs
@@ -1057,11 +1057,10 @@ namespace uAdventure.Editor
             {
                 // Folder can be created/used
                 // Does the folder exist?
-                DeleteDirectory(currentGamePath);
-                /*if (Directory.Exists(currentGamePath))
+                if (Directory.Exists(currentGamePath))
                 {
-                    // Clean the CurrentGame directory
-                }*/
+                    DeleteDirectory(currentGamePath);
+                }
                 Directory.CreateDirectory(currentGamePath);
                 create = true; 
             }
